feat: archive and roll event buffer files

+++ datanode/broker/buffered_event_source.go
@@ -5,10 +5,14 @@ import (
 	"encoding/binary"
 	"fmt"
 	"io"
-	"io/ioutil"
+	"io/fs"
 	"os"
 	"path/filepath"
+	"sort"
+	"strings"
 	"time"
+
+	"code.vegaprotocol.io/vega/datanode/utils"
 
 	"code.vegaprotocol.io/vega/core/events"
 	"code.vegaprotocol.io/vega/datanode/metrics"
@@ -24,6 +28,7 @@ type FileBufferedEventSource struct {
 	sendChannelBufferSize int
 	source                EventReceiver
 	bufferFilePath        string
+	archiveFilesPath      string
 	config                BufferedEventSourceConfig
 }
 
@@ -32,20 +37,45 @@ const (
 	numberOfSizeBytes   = 4
 )
 
-func NewBufferedEventSource(log *logging.Logger, config BufferedEventSourceConfig, source EventReceiver, bufferFilePath string) (*FileBufferedEventSource, error) {
-	err := os.RemoveAll(bufferFilePath)
+func NewBufferedEventSource(ctx context.Context, log *logging.Logger, config BufferedEventSourceConfig,
+	source EventReceiver, bufferFilesDir string,
+	archiveFilesDir string,
+) (*FileBufferedEventSource, error) {
+	err := os.RemoveAll(bufferFilesDir)
 	if err != nil {
-		return nil, fmt.Errorf("failed to remove old buffer files:%w", err)
+		return nil, fmt.Errorf("failed to remove old buffer files: %w", err)
 	}
 
-	err = os.Mkdir(bufferFilePath, os.ModePerm)
+	err = os.Mkdir(bufferFilesDir, os.ModePerm)
 	if err != nil {
-		return nil, fmt.Errorf("failed to create buffer file directory:%w", err)
+		return nil, fmt.Errorf("failed to create buffer file directory: %w", err)
 	}
 
-	files, _ := ioutil.ReadDir(bufferFilePath)
-	for _, file := range files {
-		os.Remove(filepath.Join(bufferFilePath, file.Name()))
+	if config.Archive {
+		err = os.MkdirAll(archiveFilesDir, os.ModePerm)
+		if err != nil {
+			return nil, fmt.Errorf("failed to create buffer file archive directory: %w", err)
+		}
+
+		go func() {
+			ticker := time.NewTicker(1 * time.Minute)
+			for {
+				select {
+				case <-ctx.Done():
+					return
+				case <-ticker.C:
+					err := compressUncompressedFilesInDir(archiveFilesDir)
+					if err != nil {
+						log.Errorf("failed to compress uncompressed file in archive dir: %w", err)
+					}
+
+					err = removeOldArchiveFilesIfDirectoryFull(archiveFilesDir, config.ArchiveMaximumSizeBytes)
+					if err != nil {
+						log.Errorf("failed to remove old files from full archive directory: %w", err)
+					}
+				}
+			}
+		}()
 	}
 
 	fb := &FileBufferedEventSource{
@@ -53,7 +83,8 @@ func NewBufferedEventSource(log *logging.Logger, config BufferedEventSourceConfi
 		source:             source,
 		config:             config,
 		lastBufferedSeqNum: make(chan uint64, config.MaxBufferedEvents),
-		bufferFilePath:     bufferFilePath,
+		bufferFilePath:     bufferFilesDir,
+		archiveFilesPath:   archiveFilesDir,
 	}
 
 	fb.log.Infof("Starting buffered event source with a max buffered event count of %d, and events per buffer file size %d",
@@ -188,7 +219,7 @@ func (m *FileBufferedEventSource) readEventsFromBuffer(ctx context.Context, sink
 				}
 			}
 
-			event, bufferSeqNum, read, err := readEvent(bufferFile, offset)
+			event, bufferSeqNum, read, err := readBufferedEvent(bufferFile, offset)
 			if err != nil {
 				sinkErrorCh <- fmt.Errorf("error when reading event from buffer file:%w", err)
 				return
@@ -270,14 +301,41 @@ func (m *FileBufferedEventSource) removeBufferFile(bufferFile *os.File) error {
 		return fmt.Errorf("failed to close last event buffer file:%w", err)
 	}
 
-	err = os.Remove(bufferFile.Name())
+	if m.config.Archive {
+		err = m.moveBufferFileToArchive(bufferFile.Name())
+		if err != nil {
+			return fmt.Errorf("failed to move buffer file to archive: %w", err)
+		}
+	} else {
+		err = os.Remove(bufferFile.Name())
+		if err != nil {
+			return fmt.Errorf("failed to remove event buffer file: %w", err)
+		}
+	}
+
+	return nil
+}
+
+// moveBufferFileToArchive encodes the creation time into the archive file name to ensure that the correct order
+// of files can always be determined even if the archive files are copied etc.
+func (m *FileBufferedEventSource) moveBufferFileToArchive(bufferFilePath string) error {
+	bufferFileName := filepath.Base(bufferFilePath)
+	bufferSeqSpan := strings.ReplaceAll(bufferFileName, bufferFileNamePrepend, "")
+	timeNowUtc := time.Now().UTC()
+
+	archiveFileName := fmt.Sprintf("%s-%s-%d-seqnumspan%s", bufferFileNamePrepend,
+		timeNowUtc.Format("2006-01-02-15-04-05"), timeNowUtc.UnixNano(), bufferSeqSpan)
+
+	archiveFilePath := filepath.Join(m.archiveFilesPath, archiveFileName)
+
+	err := os.Rename(bufferFilePath, archiveFilePath)
 	if err != nil {
-		return fmt.Errorf("failed to remove event buffer file:%w", err)
+		return fmt.Errorf("failed to rename file: %w", err)
 	}
 	return nil
 }
 
-func readEvent(eventFile *os.File, offset int64) (event *eventspb.BusEvent, seqNum uint64,
+func readBufferedEvent(eventFile *os.File, offset int64) (event *eventspb.BusEvent, seqNum uint64,
 	totalBytesRead uint32, err error,
 ) {
 	sizeBytes := make([]byte, numberOfSizeBytes)
@@ -322,8 +380,10 @@ func readEvent(eventFile *os.File, offset int64) (event *eventspb.BusEvent, seqN
 	return event, seqNum, totalBytesRead, nil
 }
 
+const bufferFileNamePrepend = "datanode-buffer"
+
 func (m *FileBufferedEventSource) getBufferFileName(fromSeqNum uint64, toSeqNum uint64) string {
-	return fmt.Sprintf("%s/datanode-buffer-%d-%d", m.bufferFilePath, fromSeqNum, toSeqNum)
+	return fmt.Sprintf("%s/%s-%d-%d.bevt", m.bufferFilePath, bufferFileNamePrepend, fromSeqNum, toSeqNum)
 }
 
 func (m *FileBufferedEventSource) createFile(fromSeqNum uint64, toSeqNum uint64) (*os.File, error) {
@@ -342,4 +402,80 @@ func (m *FileBufferedEventSource) openBufferFile(fromSeqNum uint64, toSeqNum uin
 		return nil, fmt.Errorf("failed to open buffer file: %s :%w", bufferFileName, err)
 	}
 	return bufferFile, nil
+}
+
+func compressUncompressedFilesInDir(dir string) error {
+	files, err := os.ReadDir(dir)
+	if err != nil {
+		return fmt.Errorf("failed to read dir: %w", err)
+	}
+
+	for _, file := range files {
+		if !file.IsDir() {
+			if !strings.HasSuffix(file.Name(), "gz") {
+				err = compressBufferedEventFile(file.Name(), dir)
+				if err != nil {
+					return fmt.Errorf("failed to compress file: %w", err)
+				}
+			}
+		}
+	}
+
+	return nil
+}
+
+func compressBufferedEventFile(bufferFileName string, archiveFilesDir string) error {
+	bufferFilePath := filepath.Join(archiveFilesDir, bufferFileName)
+	archiveFilePath := filepath.Join(archiveFilesDir, bufferFileName+".gz")
+
+	err := utils.CompressFile(bufferFilePath, archiveFilePath)
+	if err != nil {
+		return fmt.Errorf("failed to compress buffer file: %w", err)
+	}
+
+	err = os.Remove(bufferFilePath)
+	if err != nil {
+		return fmt.Errorf("failed to remove uncompressed buffer file: %w", err)
+	}
+
+	return nil
+}
+
+// removeOldArchiveFilesIfDirectoryFull intentionally uses the name of the file to figure out the relative age
+// of the file, see moveBufferFileToArchive.
+func removeOldArchiveFilesIfDirectoryFull(dir string, maximumDirSizeBytes int64) error {
+	var dirSizeBytes int64
+	var archiveFiles []fs.FileInfo
+	err := filepath.Walk(dir, func(path string, info fs.FileInfo, err error) error {
+		if !info.IsDir() {
+			dirSizeBytes += info.Size()
+			archiveFiles = append(archiveFiles, info)
+		}
+		return nil
+	})
+	if err != nil {
+		return fmt.Errorf("failed to walk directory: %w", err)
+	}
+
+	if dirSizeBytes > maximumDirSizeBytes {
+		sort.Slice(archiveFiles, func(i, j int) bool {
+			return strings.Compare(archiveFiles[i].Name(), archiveFiles[j].Name()) < 0
+		})
+
+		minimumBytesToRemove := dirSizeBytes - maximumDirSizeBytes
+
+		var bytesRemoved int64
+		for _, file := range archiveFiles {
+			err := os.Remove(filepath.Join(dir, file.Name()))
+			if err != nil {
+				return fmt.Errorf("failed to remove file: %w", err)
+			}
+			bytesRemoved += file.Size()
+			if bytesRemoved >= minimumBytesToRemove {
+				break
+			}
+		}
+	}
+
+	return nil
 }
